Fix various SVACE errors

+++ platform-abstractions/tizen/image-loaders/loader-astc.cpp
@@ -215,7 +215,7 @@ bool LoadBitmapFromAstc( const ResourceLoadingClient& client, const ImageLoader:
   size_t imageByteCount = fileSize - sizeof( AstcFileHeader );
 
   // Sanity-check the image data is not too large and that it is at less than 2 bytes per texel:
-  if( ( imageByteCount > MAX_IMAGE_DATA_SIZE ) || ( imageByteCount > ( ( width * height ) << 1 ) ) )
+  if( ( imageByteCount > MAX_IMAGE_DATA_SIZE ) || ( imageByteCount > ( ( static_cast< size_t >( width ) * height ) << 1 ) ) )
   {
     DALI_LOG_ERROR( "ASTC file has too large image-data field.\n" );
     return false;
